requirements.txt
discord.py==1.4.1
boto==2.49.0